feat: refactored the AUC logic into a separate module.
