+++ index.js
@@ -1,79 +1,23 @@
 "use strict";
+const auc = require('./auc');
 const express = require('express');
 const bodyParser = require('body-parser');
 const path = require('path');
 const port = process.env.PORT || '3002';
 
-const CPT = {
-  _FHIR_CODING_SYSTEM: 'http://www.ama-assn.org/go/cpt',
-  CARDIAC_MRI: '75561',
-  CT_HEAD_NO_CONTRAST: '70450',
-  CTA_WITH_CONTRAST: '71275',
-  LUMBAR_SPINE_CT: '72133',
-  MRA_HEAD: '70544',
-};
-
-const SNOMED = {
-  _FHIR_CODING_SYSTEM: 'http://snomed.info/sct',
-  CONGENITAL_HEART_DISEASE: '13213009',
-  HEADACHE: '25064002',
-  LOW_BACK_PAIN: '279039007',
-  OPTIC_DISC_EDEMA: '423341008',
-  TOOTHACHE: '27355003'
-};
-
-class Reasons {
-  static covers(subset, set) {
-    if (subset.size > set.size) {
-      return false;
-    }
-    for (const member of subset) {
-      if (!set.has(member)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  constructor(appropriate, notAppropriate) {
-    this.appropriate = appropriate.map(x => new Set(x));
-    this.notAppropriate = notAppropriate.map(x => new Set(x));
-  }
-
-  getRating(reasons) {
-    if (this.appropriate.filter(s => Reasons.covers(s, reasons)).length) {
-      return 'appropriate';
-    }
-    if (this.notAppropriate.filter(s => Reasons.covers(s, reasons)).length) {
-      return 'not-appropriate';
-    }
-    return 'no-guidelines-apply';
-  }
-}
-
-const cptReasons = {
-  'no-procedures-for': new Reasons([[SNOMED.TOOTHACHE]], []),
-  [CPT.CT_HEAD_NO_CONTRAST]: new Reasons([[SNOMED.HEADACHE, SNOMED.OPTIC_DISC_EDEMA]], []),
-  [CPT.MRA_HEAD]: new Reasons([], []),
-  [CPT.CTA_WITH_CONTRAST]: new Reasons([], [[SNOMED.CONGENITAL_HEART_DISEASE]]),
-  [CPT.LUMBAR_SPINE_CT]: new Reasons([], [[SNOMED.LOW_BACK_PAIN]]),
-  [CPT.CARDIAC_MRI]: new Reasons([[SNOMED.CONGENITAL_HEART_DISEASE]], []),
-};
-
 const app = express();
 app.use(bodyParser.urlencoded({extended : true}));
 app.use(bodyParser.json());
-app.get('/', function(request, response) {
-  response.redirect('/consult');
-});
+
+app.get('/', (_, response) => response.redirect('/consult'));
 
 app.get('/consult', function(request, response) {
   response.sendFile(path.join(__dirname + '/index.html'));
 });
 
 app.post('/evaluate', function(request, response) {
-  const reasons = cptReasons[request.body.procedure];
-  const rating = reasons.getRating(new Set([request.body.indication]));
+  const criterion = auc.criteria[request.body.procedure];
+  const rating = criterion.getRating(new Set([request.body.indication]));
   response.status(200).send(rating).end();
 });
